+++ src/main/kotlin/com/example/kbocombo/crawler/application/PlayerClient.kt
@@ -1,6 +1,6 @@
 package com.example.kbocombo.crawler.application
 
-import com.example.kbocombo.domain.player.Player
+import com.example.kbocombo.player.Player
 
 interface PlayerClient {
