+++ src/main/kotlin/com/example/kbocombo/player/vo/PitchingHandType.kt
@@ -1,4 +1,4 @@
-package com.example.kbocombo.domain.player.vo
+package com.example.kbocombo.player.vo
 
 enum class PitchingHandType {
     LEFT,
