+++ src/main/kotlin/com/example/kbocombo/player/vo/PlayerPosition.kt
@@ -1,4 +1,4 @@
-package com.example.kbocombo.domain.player.vo
+package com.example.kbocombo.player.vo
 
 enum class PlayerPosition {
     PITCHER,
